Refactored the api handlers to group related handlers together.  Finshed the documentation for the job and model handlers.
